Move CabalVersion and related types into a new module
